requirements.txt
asgiref==3.11.0
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0.1
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
EditorConfig==0.17.1
idna==3.11
isort==7.0.0
jsbeautifier==1.15.4
json5==0.13.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
pillow==12.2.0
platformdirs==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
pytokens==0.3.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
ruff==0.14.11
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
